feat(site/pages/AgentsPage): enable inline comments on local diffs
Extract the commenting logic (state, callbacks, annotation rendering)
from RemoteDiffPanel into a shared useDiffCommentAnnotations hook.
Wire the hook into LocalDiffPanel and thread chatInputRef through
GitPanel so local (unpushed) diffs support the same inline comment
flow as remote/PR diffs.
