+++ site/src/pages/AgentsPage/LocalDiffPanel.tsx
@@ -1,18 +1,22 @@
 import { parsePatchFiles } from "@pierre/diffs";
 import type { WorkspaceAgentRepoChanges } from "api/typesGenerated";
-import { type FC, useMemo } from "react";
-import { type DiffStyle, DiffViewer } from "./DiffViewer";
+import { type FC, type RefObject, useMemo } from "react";
+import type { ChatMessageInputRef } from "./AgentChatInput";
+import { CommentableDiffViewer } from "./CommentableDiffViewer";
+import type { DiffStyle } from "./DiffViewer";
 
 interface LocalDiffPanelProps {
 	repo: WorkspaceAgentRepoChanges;
 	isExpanded?: boolean;
 	diffStyle: DiffStyle;
+	chatInputRef?: RefObject<ChatMessageInputRef | null>;
 }
 
 export const LocalDiffPanel: FC<LocalDiffPanelProps> = ({
 	repo,
 	isExpanded,
 	diffStyle,
+	chatInputRef,
 }) => {
 	const parsedFiles = useMemo(() => {
 		const diff = repo.unified_diff;
@@ -28,11 +32,12 @@ export const LocalDiffPanel: FC<LocalDiffPanelProps> = ({
 	}, [repo.unified_diff]);
 
 	return (
-		<DiffViewer
+		<CommentableDiffViewer
 			parsedFiles={parsedFiles}
 			isExpanded={isExpanded}
 			emptyMessage="No file changes."
 			diffStyle={diffStyle}
+			chatInputRef={chatInputRef}
 		/>
 	);
 };
