+++ site/src/pages/AgentsPage/RemoteDiffPanel.tsx
@@ -1,15 +1,9 @@
-import type {
-	DiffLineAnnotation,
-	FileDiffMetadata,
-	SelectedLineRange,
-} from "@pierre/diffs";
+import type { FileDiffMetadata } from "@pierre/diffs";
 import { parsePatchFiles } from "@pierre/diffs";
 import { chatDiffContents } from "api/queries/chats";
 import type * as TypesGen from "api/typesGenerated";
-import { Button } from "components/Button/Button";
 import {
 	ArrowLeftIcon,
-	ArrowUpIcon,
 	ExternalLinkIcon,
 	GitBranchIcon,
 	GitMergeIcon,
@@ -29,15 +23,12 @@ import {
 import { useQuery } from "react-query";
 import { cn } from "utils/cn";
 import type { ChatMessageInputRef } from "./AgentChatInput";
+import { CommentableDiffViewer } from "./CommentableDiffViewer";
 import { DiffStatBadge } from "./DiffStats";
 import type { DiffStyle } from "./DiffViewer";
-import { DiffViewer } from "./DiffViewer";
-import {
-	annotationSideForBox,
-	commentBoxFromRange,
-	selectedLinesForBox,
-} from "./diffCommentSelection";
 import { parsePullRequestUrl } from "./pullRequest";
+
+export { InlinePromptInput } from "./CommentableDiffViewer";
 
 // -------------------------------------------------------------------
 // Module-level counter for cache key uniqueness
@@ -50,81 +41,6 @@ import { parsePullRequestUrl } from "./pullRequest";
  * worker pool's LRU cache persists.
  */
 let remoteDiffVersion = 0;
-
-// -------------------------------------------------------------------
-// Diff content extraction
-// -------------------------------------------------------------------
-
-/**
- * Walk the parsed hunks for a file and collect code lines that fall
- * within `startLine..endLine` on the given side. For "additions"
- * lines are matched against addition line numbers (using
- * `hunk.additionStart`); for "deletions" against deletion line
- * numbers (using `hunk.deletionStart`). Context lines that fall
- * in range are included as well.
- */
-function extractDiffContent(
-	parsedFiles: readonly FileDiffMetadata[],
-	fileName: string,
-	startLine: number,
-	endLine: number,
-	side: "additions" | "deletions",
-): string {
-	const file = parsedFiles.find((f) => f.name === fileName);
-	if (!file) return "";
-
-	const lines = side === "additions" ? file.additionLines : file.deletionLines;
-	const collected: string[] = [];
-	for (const hunk of file.hunks) {
-		let addLine = hunk.additionStart;
-		let delLine = hunk.deletionStart;
-
-		for (const block of hunk.hunkContent) {
-			if (block.type === "context") {
-				for (let i = 0; i < block.lines; i++) {
-					const ln = side === "additions" ? addLine : delLine;
-					if (ln >= startLine && ln <= endLine) {
-						const idx =
-							side === "additions"
-								? block.additionLineIndex + i
-								: block.deletionLineIndex + i;
-						if (lines[idx] != null) collected.push(lines[idx]);
-					}
-					addLine++;
-					delLine++;
-				}
-			} else {
-				// ChangeContent block.
-				if (side === "deletions") {
-					for (let i = 0; i < block.deletions; i++) {
-						if (delLine >= startLine && delLine <= endLine) {
-							const line = lines[block.deletionLineIndex + i];
-							if (line != null) collected.push(line);
-						}
-						delLine++;
-					}
-					// Addition lines in a change block still advance
-					// the addition counter.
-					addLine += block.additions;
-				} else {
-					// side === "additions"
-					// Deletion lines in a change block still advance
-					// the deletion counter.
-					delLine += block.deletions;
-					for (let i = 0; i < block.additions; i++) {
-						if (addLine >= startLine && addLine <= endLine) {
-							const line = lines[block.additionLineIndex + i];
-							if (line != null) collected.push(line);
-						}
-						addLine++;
-					}
-				}
-			}
-		}
-	}
-
-	return collected.join("\n");
-}
 
 // -------------------------------------------------------------------
 // PR state badge
@@ -166,82 +82,6 @@ const PullRequestStateBadge: FC<{
 };
 
 // -------------------------------------------------------------------
-// Inline prompt input
-// -------------------------------------------------------------------
-
-/**
- * Inline input rendered as a diff annotation under the selected
- * line(s). Supports multiline via Shift+Enter. Enter submits,
- * Escape dismisses.
- */
-export const InlinePromptInput: FC<{
-	onSubmit: (text: string) => void;
-	onCancel: () => void;
-}> = ({ onSubmit, onCancel }) => {
-	const [text, setText] = useState("");
-	const textareaRef = useRef<HTMLTextAreaElement>(null);
-
-	// Focus the textarea on mount. We use a ref callback via rAF
-	// rather than autoFocus because the component renders inside
-	// Shadow DOM where autoFocus is unreliable.
-	useEffect(() => {
-		requestAnimationFrame(() => {
-			textareaRef.current?.focus();
-		});
-	}, []);
-
-	return (
-		<div className="px-2 py-1.5">
-			<div className="rounded-lg border border-border-default/80 bg-surface-secondary/45 p-1 shadow-sm has-[textarea:focus]:ring-2 has-[textarea:focus]:ring-content-link/40">
-				<textarea
-					ref={textareaRef}
-					className="w-full resize-none border-none bg-transparent px-3 py-2 font-sans text-sm leading-5 text-content-primary placeholder:text-content-secondary outline-none ring-0 focus:outline-none focus:ring-0"
-					placeholder="Add a comment..."
-					rows={2}
-					value={text}
-					onChange={(e) => setText(e.target.value)}
-					onKeyDown={(e) => {
-						if (e.key === "Enter" && !e.shiftKey) {
-							e.preventDefault();
-							if (text.trim()) {
-								onSubmit(text.trim());
-							} else {
-								onCancel();
-							}
-						}
-						if (e.key === "Escape") {
-							e.preventDefault();
-							onCancel();
-						}
-					}}
-				/>
-				<div className="flex items-end justify-between gap-2 pl-2.5 pr-1.5 pb-1.5">
-					<span className="text-xs text-content-secondary">Esc to cancel</span>
-					<Button
-						size="icon"
-						variant="default"
-						className="size-7 rounded-full transition-colors [&>svg]:!size-4 [&>svg]:p-0"
-						disabled={!text.trim()}
-						onMouseDown={(e: React.MouseEvent) => {
-							// Prevent blur from firing before click.
-							e.preventDefault();
-						}}
-						onClick={() => {
-							if (text.trim()) {
-								onSubmit(text.trim());
-							}
-						}}
-					>
-						<ArrowUpIcon />
-						<span className="sr-only">Add to chat</span>
-					</Button>
-				</div>
-			</div>
-		</div>
-	);
-};
-
-// -------------------------------------------------------------------
 // Main component
 // -------------------------------------------------------------------
 
@@ -261,17 +101,6 @@ export const RemoteDiffPanel: FC<RemoteDiffPanelProps> = ({
 	diffStatus,
 }) => {
 	// ---------------------------------------------------------------
-	// Comment / annotation state
-	// ---------------------------------------------------------------
-	const [activeCommentBox, setActiveCommentBox] = useState<{
-		fileName: string;
-		startLine: number;
-		endLine: number;
-		side: "additions" | "deletions";
-		endSide?: "additions" | "deletions";
-	} | null>(null);
-
-	// ---------------------------------------------------------------
 	// Data fetching
 	// ---------------------------------------------------------------
 	const diffContentsQuery = useQuery({
@@ -289,7 +118,7 @@ export const RemoteDiffPanel: FC<RemoteDiffPanelProps> = ({
 
 	const parsedFiles = useMemo(() => {
 		if (!diffContent) {
-			return [];
+			return [] as FileDiffMetadata[];
 		}
 		try {
 			// The cacheKeyPrefix enables the worker pool's LRU cache
@@ -308,122 +137,9 @@ export const RemoteDiffPanel: FC<RemoteDiffPanelProps> = ({
 			);
 			return patches.flatMap((p) => p.files);
 		} catch {
-			return [];
+			return [] as FileDiffMetadata[];
 		}
 	}, [diffContent, chatId]);
-
-	// ---------------------------------------------------------------
-	// Line interaction callbacks
-	// ---------------------------------------------------------------
-	const handleLineNumberClick = useCallback(
-		(
-			fileName: string,
-			props: {
-				lineNumber: number;
-				annotationSide: "additions" | "deletions";
-			},
-		) => {
-			setActiveCommentBox({
-				fileName,
-				startLine: props.lineNumber,
-				endLine: props.lineNumber,
-				side: props.annotationSide,
-			});
-		},
-		[],
-	);
-
-	const handleLineSelected = useCallback(
-		(
-			fileName: string,
-			range: {
-				start: number;
-				end: number;
-				side?: "additions" | "deletions";
-				endSide?: "additions" | "deletions";
-			} | null,
-		) => {
-			const result = commentBoxFromRange(fileName, range);
-			if (result === "ignore") return;
-			setActiveCommentBox(result);
-		},
-		[],
-	);
-
-	// ---------------------------------------------------------------
-	// Annotation helpers
-	// ---------------------------------------------------------------
-	const getLineAnnotations = useCallback(
-		(fileName: string): DiffLineAnnotation<string>[] => {
-			if (activeCommentBox && activeCommentBox.fileName === fileName) {
-				return [
-					{
-						side: annotationSideForBox(activeCommentBox),
-						lineNumber: activeCommentBox.endLine,
-						metadata: "active-input",
-					},
-				];
-			}
-			return [];
-		},
-		[activeCommentBox],
-	);
-
-	const getSelectedLines = useCallback(
-		(fileName: string): SelectedLineRange | null => {
-			if (activeCommentBox && activeCommentBox.fileName === fileName) {
-				return selectedLinesForBox(activeCommentBox);
-			}
-			return null;
-		},
-		[activeCommentBox],
-	);
-
-	const handleCancelComment = useCallback(() => {
-		setActiveCommentBox(null);
-	}, []);
-
-	const handleSubmitComment = useCallback(
-		(text: string) => {
-			if (!activeCommentBox) return;
-			const content = extractDiffContent(
-				parsedFiles,
-				activeCommentBox.fileName,
-				activeCommentBox.startLine,
-				activeCommentBox.endLine,
-				activeCommentBox.side,
-			);
-			// Single imperative call — chip inserted atomically
-			// in one Lexical update. No rAF hack needed.
-			chatInputRef?.current?.addFileReference({
-				fileName: activeCommentBox.fileName,
-				startLine: activeCommentBox.startLine,
-				endLine: activeCommentBox.endLine,
-				content,
-			});
-			if (text.trim()) {
-				chatInputRef?.current?.insertText(text);
-			}
-			setActiveCommentBox(null);
-		},
-		[activeCommentBox, chatInputRef, parsedFiles],
-	);
-
-	const renderAnnotation = useCallback(
-		(annotation: DiffLineAnnotation<string>) => {
-			if (annotation.metadata === "active-input") {
-				if (!activeCommentBox) return null;
-				return (
-					<InlinePromptInput
-						onSubmit={handleSubmitComment}
-						onCancel={handleCancelComment}
-					/>
-				);
-			}
-			return null;
-		},
-		[activeCommentBox, handleSubmitComment, handleCancelComment],
-	);
 
 	// ---------------------------------------------------------------
 	// Scroll-to-file from chat input chip clicks
@@ -500,17 +216,13 @@ export const RemoteDiffPanel: FC<RemoteDiffPanelProps> = ({
 					</div>
 				</div>
 			)}
-			<DiffViewer
+			<CommentableDiffViewer
 				parsedFiles={parsedFiles}
 				isExpanded={isExpanded}
 				diffStyle={diffStyle}
 				isLoading={diffContentsQuery.isLoading}
 				error={diffContentsQuery.isError ? diffContentsQuery.error : undefined}
-				onLineNumberClick={handleLineNumberClick}
-				onLineSelected={handleLineSelected}
-				getLineAnnotations={getLineAnnotations}
-				getSelectedLines={getSelectedLines}
-				renderAnnotation={renderAnnotation}
+				chatInputRef={chatInputRef}
 				scrollToFile={scrollTarget}
 				onScrollToFileComplete={handleScrollComplete}
 			/>
